requirements.txt
--extra-index-url https://test.pypi.org/simple/
odd-contract-dev==0.0.22
gunicorn==20.0.4
flask==1.1.2
apscheduler==3.7.0
mysql-connector-python==8.0.23
